api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
joblib>=1.3.0
scikit-learn
xgboost
lightgbm
